requirements.txt
numpy
pandas
pyyaml
requests
tqdm
scipy
scikit-learn
matplotlib
seaborn
pybedtools
pyranges
gseapy
polars
pybigwig
torch
torchvision
lightgbm
shap
statsmodels
anndata
scanpy
sinto
